Fix/error management

+++ src/event/mode/WriteFile.cpp
@@ -13,7 +13,7 @@ WriteFile::~WriteFile() {}
 void WriteFile::Run(intptr_t offset) {
     UNUSED(offset);
     std::string content = req_.GetBody();
-    int         size    = write(polled_fd_, content.c_str(), content.size());
+    ssize_t     size    = write(polled_fd_, content.c_str(), content.size());
     if (size < 0 || static_cast<size_t>(size) != content.size())
         throw SysError("write", errno);
 }
